requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
matplotlib
nltk
pytest
httpx
python-dotenv
prometheus-client
pydantic-settings